[td] Select multiple streams in the UI
