+++ mage_ai/frontend/oracle/components/Panel/v2/index.tsx
@@ -1,0 +1,17 @@
+import { PanelStyle } from './index.style';
+
+type PanelProps = {
+  children: any;
+}
+
+function Panel({
+  children,
+}: PanelProps) {
+  return (
+    <PanelStyle>
+      {children}
+    </PanelStyle>
+  );
+}
+
+export default Panel;
